wip: rename from `webpack.js` to `build.js`
