+++ packages/bananass/src/commands/bananass-build/index.js
@@ -6,7 +6,7 @@
 // Import
 // --------------------------------------------------------------------------------
 
-import build from './webpack.js';
+import build from './build.js';
 
 // --------------------------------------------------------------------------------
 // Export
